requirements.txt
# Task 1: Data Collection and Preprocessing Dependencies

# Web Scraping
google-play-scraper>=1.2.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
